nvim: just use find_files rather than a smart git/fd picker
In the case of the nvim folder, there are submodules in the
'pack' directory that should not show up in the file picker,
but they can not be added to a .gitignore as that would break
the submodule feature. Adding them to a .ignore files does not
work for the git picker, therefore it's simple to just use a
regular file picker and maintain a tweaked .ignore file.