nvim: markdown over telekasten, hybrid treesitter
Treesitter offers too many benefits to ignore completely,
but it still does not make a lot of sense, so I'm adding
in the regular old vim syntax handling on top. This seems
to work really well, I just need to get Telekasten to play
ball with me in the markdown buffers.